Fix compiler warnings.
Xcode build has compiler warnings involving loss of precision. Remove them by adding type casts. Check results in some cases for overflow.

+++ devices/video/atirage.cpp
@@ -249,12 +249,12 @@ uint32_t ATIRage::read_reg(uint32_t reg_offset, uint32_t size) {
     }
 
     if (!offset && size == 4) { // fast path
-        return result;
+        return static_cast<uint32_t>(result);
     } else { // slow path
         if ((offset + size) > 4) {
             result |= (uint64_t)(this->regs[(reg_offset >> 2) + 1]) << 32;
         }
-        return extract_bits<uint64_t>(result, offset * 8, size * 8);
+        return static_cast<uint32_t>(extract_bits<uint64_t>(result, offset * 8, size * 8));
     }
 }
 
@@ -268,7 +268,7 @@ void ATIRage::write_reg(uint32_t reg_offset, uint32_t value, uint32_t size) {
         }
         uint64_t old_val = this->regs[reg_offset];
         insert_bits<uint64_t>(old_val, value, offset * 8, size * 8);
-        value = old_val;
+        value = static_cast<uint32_t>(old_val);
     }
 
     switch (reg_offset) {
